friend class injected hardware interface
