+++ fri/CMakeLists.txt
@@ -63,8 +63,6 @@ target_include_directories(fri
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
         $<INSTALL_INTERFACE:include>
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/src
 )
 
 # Link against dependencies
