+++ fri/src/FRIMessages.pb.c
@@ -1,7 +1,7 @@
 /* Automatically generated nanopb constant definitions */
 /* Generated by nanopb-0.2.8 at Tue Feb 14 12:42:03 2017. */
 
-#include <FRIMessages.pb.h>
+#include <fri/FRIMessages.pb.h>
